refactor: promote _advisor to public autointent.advisor package
Pipeline.fit's docstring already pointed users at autointent._advisor.run_preflight,
so the only usable entry point was behind a leading underscore. Promote the
package, keep internals private (runner.py -> _runner.py, workflows.py ->
_workflows.py), and rename the console script to match the prog= the CLI
already reports.
