+++ src/autointent/advisor/_report.py
