+++ src/autointent/custom_types/_types.py
@@ -133,7 +133,7 @@ SearchSpacePreset = Literal[
 Heavier presets explore more / larger models and take longer to run. The order
 is a cost ranking, **not** a quality ranking: a heavier preset is not strictly
 better — e.g. ``transformers-heavy`` will overfit on tiny datasets where a
-classic-* preset wins on accuracy. ``autointent._advisor.recommend`` uses this
+classic-* preset wins on accuracy. ``autointent.advisor.recommend`` uses this
 ordering to pick the heaviest preset that still fits the hardware budget,
 which is a reasonable default but not always the right choice for the data."""
